sankofa_backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.6,<5.0
psycopg2-binary>=2.9.7,<3.0
celery>=5.4,<5.5
redis>=5.0,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
Pillow>=10.4,<11.0
